+++ Lets-Run-client/letsRun/src/app/event/event-detail/event-detail.component.ts
@@ -13,17 +13,26 @@ import { DatePipe } from '@angular/common'
 import { registerLocaleData } from '@angular/common';
 import localeRu from '@angular/common/locales/ru';
 import { Datasource } from 'ngx-ui-scroll';
+import { SnackBarService } from 'src/app/services/snack-bar.service';
+import { DialogService } from 'src/app/services/DialogService';
+import * as moment from 'moment';
 
 registerLocaleData(localeRu);
 
 @Component({
   selector: 'app-event-detail',
   templateUrl: './event-detail.component.html',
-  styleUrls: ['./event-detail.component.scss']
+  styleUrls: ['./event-detail.component.scss',
+    '../../global-css/global-input.scss'
+  ]
 })
 export class EventDetailComponent implements OnInit, OnDestroy {
 
+  distances = ['Couch to 5k', '5K', '10K', 'Half Marathon', 'Marathon', 'Mud Run & fun Run', 'Trail', 'Walking'];
+
   imagePreview: string;
+
+  eventCreatedAt;
 
   event: EventModule;
   private eventSub: Subscription;
@@ -44,10 +53,12 @@ export class EventDetailComponent implements OnInit, OnDestroy {
   eventWillAttemptSubscribe: Subscription;
   containsEvent: boolean;
 
+  eventPhotoForm: FormGroup;
   eventForm: FormGroup;
   commentForm: FormGroup;
 
   edit = false;
+  editPhoto = false;
   commentInputArea = false;
 
   userId: string;
@@ -57,7 +68,13 @@ export class EventDetailComponent implements OnInit, OnDestroy {
 
   // private commentsSub: Subscription;
 
-  constructor(private eventService: EventService, private activeRoute: ActivatedRoute, private commentService: CommentService, private authService: AuthService) { }
+  constructor(private eventService: EventService,
+    private activeRoute: ActivatedRoute,
+    private commentService: CommentService,
+    private authService: AuthService,
+    private snackBarService: SnackBarService,
+    private confirm: DialogService
+  ) { }
 
   datasource = new Datasource({
     get: (index, count, success) => {
@@ -86,8 +103,12 @@ export class EventDetailComponent implements OnInit, OnDestroy {
       'location': new FormControl(null, { validators: [Validators.required] }),
       'distance': new FormControl(null, { validators: [Validators.required] }),
       'pace': new FormControl(null, { validators: [Validators.required] }),
-      image: new FormControl(null, { asyncValidators: [mimeType] }),
       'eventDate': new FormControl(null, { validators: [Validators.required] }),
+      'description': new FormControl(null, {validators: [Validators.required]}),
+    });
+
+    this.eventPhotoForm = new FormGroup({
+       image: new FormControl(null, { asyncValidators: [mimeType] })
     });
 
     this.commentForm = new FormGroup({
@@ -107,6 +128,7 @@ export class EventDetailComponent implements OnInit, OnDestroy {
       this.eventSub = this.eventService.getEventUpdate()
         .subscribe((value: { event: EventModule }) => {
           this.event = value.event;
+         this.eventCreatedAt = moment(this.event.createdAt).fromNow()
           this.checkForLikeExistance();
           this.checkForAttemptExistance();
         })
@@ -133,8 +155,8 @@ export class EventDetailComponent implements OnInit, OnDestroy {
 
   onImagePicked(event: Event) {
     const file = (event.target as HTMLInputElement).files[0];
-    this.eventForm.patchValue({ image: file });
-    this.eventForm.get('image').updateValueAndValidity();
+    this.eventPhotoForm.patchValue({ image: file });
+    this.eventPhotoForm.get('image').updateValueAndValidity();
     console.log('event.target');
     console.log(event);
 
@@ -157,21 +179,30 @@ export class EventDetailComponent implements OnInit, OnDestroy {
 
   update() {
     this.edit = !this.edit;
-
+    this.editPhoto = false;
+    this.imagePreview = null;
     let dp = new DatePipe(navigator.language);
     let p = 'y-MM-dd';
     let dtr = dp.transform(this.event.eventDate, p);
 
-    console.log(this.event.picture);
-    
+    // console.log(this.event.picture);
 
     this.eventForm.setValue({
       location: this.event.location,
       distance: this.event.distance,
       pace: this.event.pace,
-      image: this.event.picture,
-      eventDate: dtr
+      // image: this.event.picture,
+      eventDate: dtr,
+      description: this.event.description
     })
+  }
+
+
+
+  updatePhoto() {
+    this.editPhoto = !this.editPhoto;
+    this.edit = false;
+    this.imagePreview = null;
   }
 
   //******************************* */
@@ -179,8 +210,6 @@ export class EventDetailComponent implements OnInit, OnDestroy {
   //********************************* */
 
   updateEvent() {
-
-
 
     if (this.eventForm.invalid) {
       return;
@@ -193,14 +222,39 @@ export class EventDetailComponent implements OnInit, OnDestroy {
       this.eventForm.value.pace,
       this.eventForm.value.eventDate,
       this.eventForm.value.author,
-      this.eventForm.value.image
+      this.eventForm.value.description
     );
 
     this.update();
+    this.imagePreview = null;
+  }
+
+  updateEventPicture() {
+    if(this.eventPhotoForm.invalid) {
+      return;
+    }
+
+    this.eventService.uploadEventPicture(this.eventId, this.eventPhotoForm.value.image);
+
+    this.updatePhoto();
+  }
+
+  updateEventPhoto() {
+    console.log('change photo');
+    
   }
 
   deleteEvent() {
-    this.eventService.deleteEvent(this.eventId);
+    // this.eventService.deleteEvent(this.eventId);
+    this.confirm.openConfirmDialog('Are you sure want to delete the event?')
+      .afterClosed().subscribe(result => {
+        // console.log(result);
+        if (result) {
+          this.eventService.deleteEvent(this.eventId);
+          this.snackBarService.showMessageWithDuration('Event deleted', '', 3000);
+        }
+
+      })
   }
 
   //   
@@ -214,7 +268,6 @@ export class EventDetailComponent implements OnInit, OnDestroy {
 
   add_Comment() {
 
-
     if (this.commentForm.invalid) {
       return;
     }
@@ -225,32 +278,47 @@ export class EventDetailComponent implements OnInit, OnDestroy {
   }
 
   likeEvent() {
-    this.eventService.eventLikeSwitcher(this.eventId);
     this.checkForLikeExistance();
+
+    if (!this.containsLike) {
+      this.eventService.eventLikeSwitcher(this.eventId);
+      this.snackBarService.showMessageWithDuration('Event liked', '', 3000);
+    } else {
+      this.eventService.eventLikeSwitcher(this.eventId);
+      this.snackBarService.showMessageWithDuration('Event dislaiked', '', 3000);
+    }
+
   }
 
   takePart() {
+
+    if (!this.containsEvent) {
+      this.snackBarService.showMessageWithDuration('Event added to attempts', '', 3000);
+    } else {
+      this.snackBarService.showMessageWithDuration('Event removed from attempts', '', 3000);
+    }
+
     this.eventService.participateAtEvent(this.eventId);
     this.checkForAttemptExistance();
   }
 
   checkForLikeExistance() {
-    
+
     if (this.event.likes.includes(this.userId)) {
-       this.containsLike = true;
+      this.containsLike = true;
     } else {
-       this.containsLike = false;
+      this.containsLike = false;
     }
 
   }
 
   checkForAttemptExistance() {
     console.log(this.event.runners.includes(this.userId));
-    
+
     if (this.event.runners.includes(this.userId)) {
-       this.containsEvent = true;
+      this.containsEvent = true;
     } else {
-       this.containsEvent = false;
+      this.containsEvent = false;
     }
 
   }
